+++ src/app/components/demo-components.service.ts
@@ -317,6 +317,28 @@ export class SkyDemoComponentsService {
         }
       },
       {
+        name: 'Highlight',
+        icon: 'paint-brush',
+        summary: `The highlight component highlights text within DOM elements.`,
+        url: '/components/text-highlight',
+        getCodeFiles: function () {
+          return [
+            {
+              name: 'text-highlight-demo.component.html',
+              // tslint:disable-next-line
+              fileContents: require('!!raw-loader!./text-highlight/text-highlight-demo.component.html')
+            },
+            {
+              name: 'text-highlight-demo.component.ts',
+              // tslint:disable-next-line
+              fileContents: require('!!raw-loader!./text-highlight/text-highlight-demo.component.ts'),
+              componentName: 'SkyTextHighlightDemoComponent',
+              bootstrapSelector: 'sky-text-highlight-demo'
+            }
+          ];
+        }
+      },
+      {
         name: 'Key info',
         icon: 'key',
         // tslint:disable-next-line
@@ -853,7 +875,7 @@ export class SkyDemoComponentsService {
               fileContents: require('!!raw-loader!./timepicker/timepicker-demo.component.html')
             },
             {
-              name: 'tabs-demo.component.ts',
+              name: 'timepicker-demo.component.ts',
               fileContents: require('!!raw-loader!./timepicker/timepicker-demo.component.ts'),
               componentName: 'SkyTimePickerDemoComponent',
               bootstrapSelector: 'sky-timepicker-demo'