+++ src/app/components/shared/demo-page-plunker-service.ts
@@ -61,18 +61,22 @@ export class SkyDemoPagePlunkerService {
 
     'app': '.',
 
-    '@angular/core': 'npm:@angular/core@2.4.5/bundles/core.umd.js',
-    '@angular/common': 'npm:@angular/common@2.4.5/bundles/common.umd.js',
-    '@angular/compiler': 'npm:@angular/compiler@2.4.5/bundles/compiler.umd.js',
-    '@angular/platform-browser': 'npm:@angular/platform-browser@2.4.5/bundles/platform-browser.umd.js',
-    '@angular/platform-browser-dynamic': 'npm:@angular/platform-browser-dynamic@2.4.5/bundles/platform-browser-dynamic.umd.js',
-    '@angular/http': 'npm:@angular/http@2.4.5/bundles/http.umd.js',
-    '@angular/router': 'npm:@angular/router@3.4.5/bundles/router.umd.js',
-    '@angular/forms': 'npm:@angular/forms@2.4.5/bundles/forms.umd.js',
+    '@angular/core': 'npm:@angular/core@4.2.5/bundles/core.umd.js',
+    '@angular/common': 'npm:@angular/common@4.2.5/bundles/common.umd.js',
+    '@angular/compiler': 'npm:@angular/compiler@4.2.5/bundles/compiler.umd.js',
+    '@angular/platform-browser': 'npm:@angular/platform-browser@4.2.5/bundles/platform-browser.umd.js',
+    '@angular/platform-browser-dynamic': 'npm:@angular/platform-browser-dynamic@4.2.5/bundles/platform-browser-dynamic.umd.js',
+    '@angular/http': 'npm:@angular/http@4.2.5/bundles/http.umd.js',
+    '@angular/router': 'npm:@angular/router@4.2.5/bundles/router.umd.js',
+    '@angular/forms': 'npm:@angular/forms@4.2.5/bundles/forms.umd.js',
+    '@angular/animations': 'npm:@angular/animations@4.2.5/bundles/animations.umd.js',
+    '@angular/platform-browser/animations': 'npm:@angular/platform-browser@4.2.5/bundles/platform-browser-animations.umd.js',
+    '@angular/animations/browser': 'npm:@angular/animations@4.2.5/bundles/animations-browser.umd.js',
+    'tslib': 'npm:tslib@1.6.1',
 
-    'rxjs': 'npm:rxjs@5.1.0',
-    'typescript': 'npm:typescript@2.0.2/lib/typescript.js',
-    '@blackbaud/skyux/dist/core': 'npm:@blackbaud/skyux@2.0.0-beta.31/dist/bundles/core.umd.js',
+    'rxjs': 'npm:rxjs',
+    'typescript': 'npm:typescript@2.2.1/lib/typescript.js',
+    '@blackbaud/skyux/dist/core': 'npm:@blackbaud/skyux/dist/bundles/core.umd.js',
 
     'moment': 'npm:moment/moment.js',