+++ src/modules/text-highlight/text-highlight.module.ts
@@ -1,0 +1,16 @@
+import { NgModule } from '@angular/core';
+import { SkyTextHighlightDirective } from './text-highlight.directive';
+import { MutationObserverService } from '../mutation/mutation-observer-service';
+
+@NgModule({
+  declarations: [
+    SkyTextHighlightDirective
+  ],
+  exports: [
+    SkyTextHighlightDirective
+  ],
+  providers: [
+    MutationObserverService
+  ]
+})
+export class SkyTextHighlightModule { }